Move StatusRequestsHandler to request_handlers.py
